refactor: src/utils로 로직 분리
